Add symlink to correctly named file
